quick fix about the products and user models
